+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Simplified configuration. Almost all configuration options now have sane defaults
 - Deprecated `bPosRand` and `bSclRand`
 - Removed z offset and id prefix from the UI and moved it to the config file
+- No longer write `REGN`, `NAM5` to `CELL` records
 
 ### Removed
 - Removed configuration options `sName`, `sRecType` and `sScript`